Pushed 3 functions down to Lemmas
